+++ src/main/java/org/mifos/connector/ams/fineract/camel/config/CamelContextConfig.java
@@ -1,19 +1,28 @@
 package org.mifos.connector.ams.fineract.camel.config;
 
 import java.util.HashMap;
+import lombok.RequiredArgsConstructor;
 import org.apache.camel.CamelContext;
+import org.apache.camel.component.http.HttpComponent;
 import org.apache.camel.spi.RestConfiguration;
 import org.apache.camel.spring.boot.CamelContextConfiguration;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.info.BuildProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 /** Configuration class for camel context. */
 @Configuration
+@RequiredArgsConstructor
 public class CamelContextConfig {
 
     @Value("${camel.server-port}")
     private int serverPort;
+
+    private final BuildProperties buildProperties;
+
+    @Value("${spring.application.name:ph-ee-connector-ams-fineract}")
+    private String applicationName;
 
     /**
      * Configures the camel context to be used.
@@ -40,6 +49,8 @@ public class CamelContextConfig {
                 rest.setDataFormatProperties(new HashMap<>());
                 rest.getDataFormatProperties().put("prettyPrint", "true");
                 rest.setScheme("http");
+                configureHttpComponent(camelContext, "http");
+                configureHttpComponent(camelContext, "https");
             }
 
             @Override
@@ -47,5 +58,19 @@ public class CamelContextConfig {
                 // empty
             }
         };
+    }
+
+    /**
+     * Configures the HTTP component with a user agent.
+     *
+     * @param camelContext
+     *            the Camel context to configure
+     * @param scheme
+     *            the scheme to configure (http or https)
+     */
+    private void configureHttpComponent(CamelContext camelContext, String scheme) {
+        HttpComponent httpComponent = camelContext.getComponent(scheme, HttpComponent.class);
+        String userAgent = String.format("%s/%s", applicationName, buildProperties.getVersion());
+        httpComponent.setHttpClientConfigurer(clientBuilder -> clientBuilder.setUserAgent(userAgent));
     }
 }